# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Noroff EP1 E-commerce API Documentation",
    "description": "API documentation for the project",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter JWT token in the format: Bearer <token>"
    }
  },
  "paths": {
...
    "/admin/orders": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Endpoint to get all orders",
        "parameters": [
          {
            "name": "accept",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Orders retrieved successfully"
                    },
                    "orders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/Order"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/admin/orders/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Endpoint to get order by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Order ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Order retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Order retrieved successfully"
                    },
                    "order": {
                      "$ref": "#/definitions/Order"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Order not found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 404
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Order not found"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/orders": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get all orders for the logged in user",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Orders found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Orders found"
                    },
                    "orders": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "status": {
                            "type": "string",
                            "example": "In progress"
                          },
                          "order_number": {
                            "type": "string",
                            "example": "abcdefgh"
                          },
                          "discount": {
                            "type": "number",
                            "example": 0
                          },
                          "membership_status": {
                            "type": "string",
                            "example": "Bronze"
                          },
                          "createdAt": {
                            "type": "string",
                            "example": "2023-01-01T00:00:00.000Z"
                          },
                          "updatedAt": {
                            "type": "string",
                            "example": "2023-01-01T00:00:00.000Z"
                          },
                          "OrderItems": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "number",
                                  "example": 1
                                },
                                "product_id": {
                                  "type": "number",
                                  "example": 1
                                },
                                "quantity": {
                                  "type": "number",
                                  "example": 2
                                },
                                "unit_price": {
                                  "type": "number",
                                  "example": 100
                                },
                                "Product": {
                                  "type": "object",
                                  "properties": {
                                    "id": {
                                      "type": "number",
                                      "example": 1
                                    },
                                    "name": {
                                      "type": "string",
                                      "example": "Product Name"
                                    },
                                    "description": {
                                      "type": "string",
                                      "example": "Product Description"
                                    },
                                    "price": {
                                      "type": "number",
                                      "example": 100
                                    },
                                    "category_id": {
                                      "type": "number",
                                      "example": 1
                                    },
                                    "brand_id": {
                                      "type": "number",
                                      "example": 1
                                    },
                                    "quantity": {
                                      "type": "number",
                                      "example": 10
                                    },
                                    "createdAt": {
                                      "type": "string",
                                      "example": "2023-01-01T00:00:00.000Z"
                                    },
                                    "updatedAt": {
                                      "type": "string",
                                      "example": "2023-01-01T00:00:00.000Z"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "No orders found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 404
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "No orders found"
                    },
                    "orders": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 500
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Error message"
                    },
                    "orders": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get an order by ID for the logged in user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Order ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Order found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Order found"
                    },
                    "order": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "status": {
                          "type": "string",
                          "example": "In progress"
                        },
                        "order_number": {
                          "type": "string",
                          "example": "abcdefgh"
                        },
                        "discount": {
                          "type": "number",
                          "example": 0
                        },
                        "membership_status": {
                          "type": "string",
                          "example": "Bronze"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2023-01-01T00:00:00.000Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2023-01-01T00:00:00.000Z"
                        },
                        "OrderItems": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number",
                                "example": 1
                              },
                              "product_id": {
                                "type": "number",
                                "example": 1
                              },
                              "quantity": {
                                "type": "number",
                                "example": 2
                              },
                              "unit_price": {
                                "type": "number",
                                "example": 100
                              },
                              "Product": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "number",
                                    "example": 1
                                  },
                                  "name": {
                                    "type": "string",
                                    "example": "Product Name"
                                  },
                                  "description": {
                                    "type": "string",
                                    "example": "Product Description"
                                  },
                                  "price": {
                                    "type": "number",
                                    "example": 100
                                  },
                                  "category_id": {
                                    "type": "number",
                                    "example": 1
                                  },
                                  "brand_id": {
                                    "type": "number",
                                    "example": 1
                                  },
                                  "quantity": {
                                    "type": "number",
                                    "example": 10
                                  },
                                  "createdAt": {
                                    "type": "string",
                                    "example": "2023-01-01T00:00:00.000Z"
                                  },
                                  "updatedAt": {
                                    "type": "string",
                                    "example": "2023-01-01T00:00:00.000Z"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Order not found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 404
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Order not found"
                    },
                    "order": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 500
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Error message"
                    },
                    "order": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
...
    "Order": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "In progress"
        },
        "order_number": {
          "type": "string",
          "example": "abcdefgh"
        },
        "discount": {
          "type": "number",
          "example": 0
        }
      }
    }
  }
}